+++ src/revert/function_account_birth_date_update.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+DROP FUNCTION vibetype.account_birth_date_update(DATE);
+
+COMMIT;