+++ src/revert/function_account_location_update.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+DROP FUNCTION vibetype.account_location_update(DOUBLE PRECISION, DOUBLE PRECISION);
+
+COMMIT;